Ensure item destruction doesn't read freed obj
This one is really tricky and someone should sanity check this before it
gets merged...

When items get destroyed by a wand of lightning, it builds up a stack of
frames to check as items may explode destroying other items. This is a
little tricky though, and the current item may be destroyed as part of
this routine.

I modified the guard to also ensure the current frame is checked, not
just the subsequent frames, and haven't since seen the address sanitizer
tripping on this.